+++ src/components/ids-data-grid/demos/empty-message-custom-thru-slot.ts
@@ -1,19 +1,12 @@
 import type IdsDataGrid from '../ids-data-grid';
-import type IdsPopupMenu from '../../ids-popup-menu/ids-popup-menu';
-import type IdsMenuItem from '../../ids-menu/ids-menu-item';
 import '../ids-data-grid';
 import type { IdsDataGridColumn } from '../ids-data-grid-column';
 import '../../ids-container/ids-container';
 import productsJSON from '../../../assets/data/products.json';
 
 // Example for populating the DataGrid
-const dataGrid = document.querySelector<IdsDataGrid>('#data-grid-save-settings')!;
-const rowHeightMenu = document.querySelector<IdsPopupMenu>('#row-height-menu')!;
-
-// Change row height with popup menu
-rowHeightMenu?.addEventListener('selected', (e: Event) => {
-  dataGrid.rowHeight = (e.target as IdsMenuItem).value as string;
-});
+const dataGrid = document.querySelector<IdsDataGrid>('#data-grid-empty-message-custom-thru-slot')!;
+const showEmptyData = true;
 
 (async function init() {
   const columns: IdsDataGridColumn[] = [];
@@ -26,16 +19,6 @@ rowHeightMenu?.addEventListener('selected', (e: Event) => {
     resizable: false,
     formatter: dataGrid.formatters.selectionCheckbox,
     align: 'center'
-  });
-  columns.push({
-    id: 'id',
-    name: 'ID',
-    field: 'id',
-    width: 80,
-    resizable: true,
-    reorderable: true,
-    sortable: true,
-    formatter: dataGrid.formatters.text,
   });
   columns.push({
     id: 'color',
@@ -79,17 +62,6 @@ rowHeightMenu?.addEventListener('selected', (e: Event) => {
     }]
   });
   columns.push({
-    id: 'inStock',
-    name: 'In Stock',
-    field: 'inStock',
-    sortable: true,
-    resizable: true,
-    reorderable: true,
-    align: 'center',
-    formatter: dataGrid.formatters.text,
-    filterType: dataGrid.filters.checkbox
-  });
-  columns.push({
     id: 'unitPrice',
     name: 'Unit Price',
     field: 'unitPrice',
@@ -117,8 +89,13 @@ rowHeightMenu?.addEventListener('selected', (e: Event) => {
   const setData = async () => {
     const res = await fetch(url);
     const data = await res.json();
-    dataGrid.data = data;
-    dataGrid.pageTotal = data.length;
+
+    if (showEmptyData) {
+      dataGrid.data = [];
+    } else {
+      dataGrid.data = data;
+      dataGrid.pageTotal = data.length;
+    }
   };
   setData();
 }());